requirements.txt
absl-py==0.9.0
backcall==0.1.0
beautifulsoup4==4.9.0
bs4==0.0.1
cachetools==4.2.1
certifi==2020.12.5
chardet==3.0.4
colorlog==4.1.0
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
future==0.18.2
google-auth==1.27.0
google-auth-oauthlib==0.4.2
greenlet==0.4.15
grpcio==1.28.1
idna==2.8
importlib-metadata==3.4.0
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.0
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.2.1
msgpack==1.0.0
numpy==1.16.3
oauthlib==3.1.0
opencv-python==4.5.1.48
pandas==1.0.3
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prompt-toolkit==3.0.16
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pynvim==0.4.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.7.1
scipy==1.5.4
seaborn==0.10.1
six==1.14.0
soupsieve==2.0
tensorboard==1.14.0
tensorboard-plugin-wit==1.8.0
torch==1.5.0
torchvision==0.6.0
tqdm==4.46.0
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.25.9
wcwidth==0.1.9
Werkzeug==1.0.1
zipp==3.1.0
